refactor: restructure the folder, file name and make the constant px to theme css variable
