+++ wishlist/frontend/src/app/layout.tsx
@@ -1,10 +1,10 @@
-import HeaderBar from '@/components/header-bar'
 import { Metadata } from 'next'
-import './globals.css'
-import Content from '@/components/content'
+import '@/styles/globals.css'
 import { ReactNode } from 'react'
 import axios from 'axios'
 import { Poppins } from 'next/font/google'
+import HeaderBar from '@/components/layout/HeaderBar'
+import LeftPanelContainer from '@/components/layout/LeftPanelContainer'
 
 axios.defaults.baseURL = `http://${process.env.BACKEND_HOST}:${process.env.BACKEND_PORT}`
 
@@ -22,10 +22,9 @@ export default function RootLayout({ children }: { children: ReactNode }) {
   return (
     <html lang="en" className={poppins.className} suppressHydrationWarning>
       <body className="h-screen min-w-[400px] overflow-y-hidden">
-        {/* The height of header-bar is 60px, thus the container should -60px */}
-        <HeaderBar></HeaderBar>
-        <div className="h-[calc(100%-60px)]">
-          <Content>{children}</Content>
+        <HeaderBar />
+        <div className="h-[calc(100%-var(--header-height))]">
+          <LeftPanelContainer>{children}</LeftPanelContainer>
         </div>
       </body>
     </html>
