+++ wishlist/frontend/src/components/common/Icon.tsx
