+++ wishlist/frontend/src/components/common/tag/ColorTag.tsx
