+++ wishlist/frontend/src/components/common/tag/DotTag.tsx
