+++ wishlist/frontend/src/components/common/tag/IconTag.tsx
@@ -2,7 +2,7 @@ import { COLOR } from '@/enum/color'
 import { ICON } from '@/enum/icon'
 import { FaClock } from 'react-icons/fa6'
 import * as TSP from 'ts-pattern'
-import Icon from '../icon'
+import Icon from '../Icon'
 
 type IconTagProps = {
   icon: ICON
