+++ wishlist/frontend/src/components/layout/HeaderBar.tsx
@@ -5,9 +5,9 @@ export default async function HeaderBar() {
   await connection()
   const color = process.env.A_COLOR
   return (
-    <div
+    <aside
       data-testid="header-bar"
-      className="h-[60px] flex items-center justify-between border-b-2 border-slate-600">
+      className="h-[var(--header-height)] flex items-center justify-between border-b-2 border-slate-600">
       <div data-testid="header-bar-title" className="pl-7 text-xl">
         <Image
           className="inline object-center w-8 h-8"
@@ -19,7 +19,7 @@ export default async function HeaderBar() {
         <span className="pl-4">Wishlist</span>
       </div>
       <div>Process environment : {color}</div>
-      <div data-testid="header-bar-username" className="pr-7 text-xl">
+      <div data-testid="header-bar-user" className="pr-7 text-xl">
         <span className="pr-4">Lilypop</span>
         <Image
           className="inline rounded-full w-10 h-10"
@@ -29,6 +29,6 @@ export default async function HeaderBar() {
           height="40"
         />
       </div>
-    </div>
+    </aside>
   )
 }
