+++ wishlist/frontend/src/components/layout/LeftPanelContainer.tsx
@@ -1,0 +1,24 @@
+'use client'
+import { ReactNode } from 'react'
+import { leftPanelVisibleAtom } from '@/atoms/atoms'
+import { useAtomValue } from 'jotai'
+import LeftPanel from './LeftPanel'
+import clsx from 'clsx'
+
+export default function LeftPanelContainer({ children }: { children: ReactNode }) {
+  const leftPanelVisible = useAtomValue(leftPanelVisibleAtom)
+
+  return (
+    <div className="relative w-full h-full overflow-x-hidden">
+      <LeftPanel />
+      <div
+        data-testid="content"
+        className={clsx(
+          'absolute left-[50px] shrink h-full duration-500 ease-out transition-all',
+          leftPanelVisible ? 'left-panel-expanded' : 'left-panel-collapsed'
+        )}>
+        {children}
+      </div>
+    </div>
+  )
+}
