+++ wishlist/frontend/src/components/todolist/TaskCard.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { currentTaskAtom } from '@/atoms/atoms'
-import ColorTag from '@/components/tag/color-tag'
-import IconTag from '@/components/tag/icon-tag'
+import ColorTag from '@/components/common/tag/ColorTag'
+import IconTag from '@/components/common/tag/IconTag'
 import { COLOR } from '@/enum/color'
 import { ICON } from '@/enum/icon'
 import { EMERGENCY, IMPORTANT, Task } from '@/types/tasks/task'
